Moved InstrumentQCFailuresViewlet to viewlets package
