+++ bika/lims/browser/instrument.zcml
@@ -76,15 +76,4 @@
       layer="bika.lims.interfaces.IBikaLIMS"
     />
 
-    <!-- Show Instrument QC failures viewlet -->
-    <browser:viewlet
-        for="*"
-        name="bika.lims.instrument_qc_failures_viewlet"
-        class="bika.lims.browser.instrument.InstrumentQCFailuresViewlet"
-        manager="plone.app.layout.viewlets.interfaces.IAboveContent"
-        template="templates/instrument_qc_failures_viewlet.pt"
-        permission="zope2.View"
-        layer="bika.lims.interfaces.IBikaLIMS"
-    />
-
 </configure>
