+++ bika/lims/browser/viewlets/__init__.py
@@ -11,3 +11,4 @@ from document_actions import DocumentActionsViewlet  # noqa
 from new_versions import NewVersionsViewlet  # noqa
 from path_bar import PathBarViewlet  # noqa
 from authenticator import AuthenticatorViewlet  # noqa
+from instruments import InstrumentQCFailuresViewlet  # noqa
