+++ bika/lims/browser/viewlets/configure.zcml
@@ -116,4 +116,15 @@
       permission="zope2.View"
       />
 
+    <!-- Show Instrument QC failures viewlet -->
+    <browser:viewlet
+        for="*"
+        name="bika.lims.instrument_qc_failures_viewlet"
+        class="bika.lims.browser.viewlets.InstrumentQCFailuresViewlet"
+        manager="plone.app.layout.viewlets.interfaces.IAboveContent"
+        template="templates/instrument_qc_failures_viewlet.pt"
+        permission="zope2.View"
+        layer="bika.lims.interfaces.IBikaLIMS"
+    />
+
 </configure>
